pyproject.toml
[project]
name = "post-converter"
version = "0.1.0"
description = "CLI tool to convert Markdown blog posts to Python scripts and Jupyter Notebooks"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer>=0.9.0",
    "python-frontmatter>=1.0.0",
    "nbformat>=5.9.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "markdown-it-py>=3.0.0",
]

[project.scripts]
post-converter = "post_converter.main:app"

[tool.ruff]
line-length = 88
target-version = "py39"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
